Added Transitions
Took out transitions during the refactor, so added them back in.
Transitions exist for node click but not for simple redrawing of the
graph (since we don’t want a transition from the origin at this point).
However, transition for node removal is a bit strange. It would be
better if it behaved more like the node click, so this still needs a
bit more work.